[BEAM-7372] cleanup py2 and py35 codepath from apache_beam/typehints * usage of future package * usage of past package * unnecessary builtins import * workaround for python 2 and 3.5 * test files only for python 3
