+++ sdks/python/apache_beam/typehints/native_type_compatibility_test.py
@@ -19,8 +19,6 @@
 
 # pytype: skip-file
 
-from __future__ import absolute_import
-
 import sys
 import typing
 import unittest
@@ -70,7 +68,6 @@ class NativeTypeCompatibilityTest(unittest.TestCase):
             bytes, typing.Union[int, bytes, float]]]],
          typehints.Tuple[bytes, typehints.List[typehints.Tuple[
              bytes, typehints.Union[int, bytes, float]]]]),
-        # TODO(BEAM-7713): This case seems to fail on Py3.5.2 but not 3.5.4.
         ('arbitrary-length tuple', typing.Tuple[int, ...],
          typehints.Tuple[int, ...]),
         ('flat alias', _TestFlatAlias, typehints.Tuple[bytes, float]),  # type: ignore[misc]
