+++ sdks/python/apache_beam/typehints/row_type.py
@@ -17,8 +17,6 @@
 
 # pytype: skip-file
 
-from __future__ import absolute_import
-
 from apache_beam.typehints import typehints
