+++ sdks/python/apache_beam/typehints/typecheck.py
@@ -22,14 +22,10 @@ For internal use only; no backwards-compatibility guarantees.
 
 # pytype: skip-file
 
-from __future__ import absolute_import
-
 import collections
 import inspect
+import sys
 import types
-
-from future.utils import raise_with_traceback
-from past.builtins import unicode
 
 from apache_beam import pipeline
 from apache_beam.pvalue import TaggedOutput
@@ -93,7 +89,8 @@ class OutputCheckWrapperDoFn(AbstractDoFnWrapper):
       error_msg = (
           'Runtime type violation detected within ParDo(%s): '
           '%s' % (self.full_label, e))
-      raise_with_traceback(TypeCheckError(error_msg))
+      _, _, tb = sys.exc_info()
+      raise TypeCheckError(error_msg).with_traceback(tb)
     else:
       return self._check_type(result)
 
@@ -102,7 +99,7 @@ class OutputCheckWrapperDoFn(AbstractDoFnWrapper):
     if output is None:
       return output
 
-    elif isinstance(output, (dict, bytes, str, unicode)):
+    elif isinstance(output, (dict, bytes, str)):
       object_type = type(output).__name__
       raise TypeCheckError(
           'Returning a %s from a ParDo or FlatMap is '
@@ -184,13 +181,15 @@ class TypeCheckWrapperDoFn(AbstractDoFnWrapper):
     try:
       check_constraint(type_constraint, datum)
     except CompositeTypeHintError as e:
-      raise_with_traceback(TypeCheckError(e.args[0]))
+      _, _, tb = sys.exc_info()
+      raise TypeCheckError(e.args[0]).with_traceback(tb)
     except SimpleTypeHintError:
       error_msg = (
           "According to type-hint expected %s should be of type %s. "
           "Instead, received '%s', an instance of type %s." %
           (datum_type, type_constraint, datum, type(datum)))
-      raise_with_traceback(TypeCheckError(error_msg))
+      _, _, tb = sys.exc_info()
+      raise TypeCheckError(error_msg).with_traceback(tb)
 
 
 class TypeCheckCombineFn(core.CombineFn):
@@ -220,7 +219,8 @@ class TypeCheckCombineFn(core.CombineFn):
         error_msg = (
             'Runtime type violation detected within %s: '
             '%s' % (self._label, e))
-        raise_with_traceback(TypeCheckError(error_msg))
+        _, _, tb = sys.exc_info()
+        raise TypeCheckError(error_msg).with_traceback(tb)
     return self._combinefn.add_input(accumulator, element, *args, **kwargs)
 
   def merge_accumulators(self, accumulators, *args, **kwargs):
@@ -239,7 +239,8 @@ class TypeCheckCombineFn(core.CombineFn):
         error_msg = (
             'Runtime type violation detected within %s: '
             '%s' % (self._label, e))
-        raise_with_traceback(TypeCheckError(error_msg))
+        _, _, tb = sys.exc_info()
+        raise TypeCheckError(error_msg).with_traceback(tb)
     return result
 
   def teardown(self, *args, **kwargs):
